+++ test/barcode.jl
@@ -20,7 +20,7 @@ using ZXingCPP:
     end
     expected = "96385074"
 
-    iv = ImageView(data, Int32(data_len), Int32(1), ZXing_ImageFormat_Lum)
+    iv = ImageView(data, data_len, 1, ZXing_ImageFormat_Lum)
     ro = ReaderOptions(; binarizer=ZXing_Binarizer_BoolCast)
     bcs = read_barcodes(iv, ro)
     bc = bcs[1]
